% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPC_functions.R
\name{MPC}
\alias{MPC}
\title{Metapopulation capacity (MPC) core function}
\usage{
MPC(pa, mdist, alpha, dispfun, ex, self, evec, symmetric, lower, savememory)
}
\arguments{
\item{pa}{A numeric vector of areas of habitat patches 1-n}

\item{mdist}{A square matrix of pairwise distances between habitat patches 1-n}

\item{alpha}{A species-specific dispersal distance (e.g. average gap-crossing distance), default=317m}

\item{dispfun}{A species-specific dispersal survival function, currently 3 options can be chosen by parameter dispfop: "negex", "linear" and "log-sech"}

\item{ex}{A stochasticity parameter: if x >1, it becomes very unlikely for populations to go extinct after a critical patch size has been reached. If x <1, there exists no critical patch size and populations can even go extinct if they are large (Hanski 1994)}

\item{self}{Logical: should self-colonization of patches be modeled (default, TRUE) or not (FALSE)}

\item{evec}{Logical: should the dominant eigenvalue associated eigenvector be returned (default, TRUE) or not (FALSE)}

\item{symmetric}{Logical: is the pairwise distance matrix symmetric (TRUE) or not (FALSE, default). In case symmetric = TRUE, eigenvalue calculations are faster because only the lower triangle will be considered. If symmetric is not specified, the eigenvalue function will use the whole matrix as input.}

\item{lower}{Logical: In case the symmetric argument is TRUE: which triangle of the matrix should be used for calculation, the lower left (TRUE, default) or upper right (FALSE)}
}
\value{
An 'MPC' object: a list containing a numeric MPC value (mpc), metapopulation density (mpcdens), the MPC-associated squared dominant eigenvector indicating patch importance (pimport) and the patch IDs in case they are indicated (pid)
}
\description{
Calculates MPC for a given set of habitat patches. The function relies on the 'sf' and 'stars'packages
}
