\name{MPC_linreg}
\alias{MPC_linreg}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
MPC for linear reg
}
\description{
scale the posterior for linear reg
}
\usage{
MPC_linreg(nn, data, theta_boot, data_boot, alpha, M_samp, B_resamp)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{nn}{
	a
}
  \item{data}{
	a
}
  \item{theta_boot}{
a
}
  \item{data_boot}{
a
}
  \item{alpha}{
	a
}
  \item{M_samp}{
	a
}
  \item{B_resamp}{
	a
}
}
\value{
a
}

\examples{
\dontrun{
a
}
}

