% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPC_functions.R
\name{MPC_series}
\alias{MPC_series}
\alias{MPCser}
\title{calculate a series of Metapopulation Capacity (MPC) - based metrics}
\usage{
MPCser(
  roi = NULL,
  MPCbasl = NULL,
  x = NULL,
  pa = NULL,
  mdist = NULL,
  alpha = 317,
  dispfop = "log-sech",
  ex = 0.5,
  self = TRUE,
  evec = TRUE,
  symmetric = FALSE,
  lower = TRUE,
  distfun = sf::st_distance,
  areafun = sf::st_area,
  savememory = TRUE
)
}
\arguments{
\item{roi}{shapefile of region of interest}

\item{MPCbasl}{numeric value of baseline MPCvalue}
}
\value{
a list with MPC based metrics: MPCraw = raw MPC value, MPCev = eigenvector for each patch ID, MPCdens=MPC density, MPCrmax=MPC relative to maximum potential.
}
\description{
it uses the same arguments as the MPC_fun with two additional ones: roi and a baseline MPCvalue
}
