\name{MPC_varmix}
\alias{MPC_varmix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
MPC for variational posterior of a normal mixture model
}
\description{
calibrate the variational posterior for a normal mixture model
}
\usage{
MPC_varmix(nn, data, data_boot, alpha, B_resamp)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{nn}{
	a
}
  \item{data}{
	a
}
  \item{data_boot}{
a
}
  \item{alpha}{
	a
}
  \item{B_resamp}{
	a
}
}
\value{
a
}

\examples{
\dontrun{
a
}
}

