% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPC_functions.R
\name{MPC_fun}
\alias{MPC_fun}
\title{Metapopulation capacity (MPC) wrapper function}
\usage{
MPC_fun(
  x = NULL,
  pa = NULL,
  mdist = NULL,
  alpha = 317,
  dispfop = "log-sech",
  ex = 0.5,
  self = TRUE,
  evec = TRUE,
  symmetric = FALSE,
  lower = TRUE,
  areafun = sf::st_area,
  distfun = sf::st_distance,
  savememory = TRUE
)
}
\arguments{
\item{x}{A raster object where habitat patches are set to 1 and matrix area is set to NA OR a shapefile object containing habitat patches}

\item{pa}{A numeric vector of areas of habitat patches 1-n}

\item{mdist}{A square matrix of pairwise distances between habitat patches 1-n}

\item{alpha}{A species-specific dispersal distance (e.g. average gap-crossing distance), default=317m}

\item{dispfop}{An option to choose the dispersal survival function (dispfun parameter), currently 3 options implemented: "negex", "linear" and "log-sech"}

\item{ex}{A stochasticity parameter: if x >1, it becomes very unlikely for populations to go extinct after a critical patch size has been reached. If x <1, there exists no critical patch size and populations can even go extinct if they are large (Hanski 1994)}

\item{self}{Logical: should self-colonization of patches be modeled (default, TRUE) or not (FALSE)}

\item{evec}{Logical: should the dominant eigenvalue associated eigenvector be returned (default, TRUE) or not (FALSE)}

\item{symmetric}{Logical: is the pairwise distance matrix symmetric (TRUE) or not (FALSE, default). In case symmetric = TRUE, eigenvalue calculations are faster because only the lower triangle will be considered. If symmetric is not specified, the eigenvalue function will use the whole matrix as input.}

\item{lower}{Logical: In case the symmetric argument is TRUE: which triangle of the matrix should be used for calculation, the lower left (TRUE, default) or upper right (FALSE)}

\item{areafun}{A function defining how patch areas are calculated in case the input is a raster or shapefile (default: sf::st_area)}

\item{distfun}{A function defining how interpatch distances are calculated in case the input is a raster or shapefile (default: euclidean, nearest, edge-to-edge: sf::st_distance)}
}
\value{
the 'MPC' object created by the function 'MPC'
}
\description{
Calculates MPC for a given set of habitat patches. The function relies on the 'sf' and 'stars'packages
}
